Extract buildUri function for reuse between B2B and consumer; address list and map types properly
